Move the tests to filehelpersTest class
